pyproject.toml
[tool.poetry]
name = "imdb_ratings"
version = "0.1.0"
description = ""
authors = ["Xingyu Zhang <xingyuz123@gmail.com>"]
readme = "README.md"
packages = [
    {include = "imdb_ratings"}
]

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.18.0"
numpy = "^2.2.1"
bs4 = "^0.0.2"
pydantic = "^2.10.4"
supabase = "^2.11.0"
python-dotenv = "^1.0.1"
urllib3 = "^2.3.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
